Move FileType enum from sql module to logical_plan module
